Stop defaulting to create a new logger for viewer utilities
It makes it a bit difficult to follow what module is in charge of
providing defaults if every one of them is overly protective and knows
how instantiate their loggers.

Defining clearer boundaries for the individual pieces allows for safer
refactoring, and should something fail, the parameter invariant should
be triggered to get a good bug report.