Require bundleDir to always be present
The `bundleDir` value can come in via only two ways:

1. From the **plugin instance** getting it from
   `this.compiler.outputPath`

2. From the **CLI** arguments or inferring it from the bundle stat file
   location

It seems like there shouldn't be any scenario where the `bundleDir`
should be `null` or `undefined` after these steps. If I haven't
misunderstood, the `this.compiler.outputPath` comes from `output.path`
webpack config value which in turn is required by webpack to be set.

So it should be fairly safe to remove the overly protective check from
the code.