Allow reports to be analyzed even with incomplete data
This partially addresses #29 by allowing the report to be generated at
all in the first place. Now only the parts of the bundle that failed to
be analyzed don't contain sizes for Gzip or parsed parts, but will still
render successfully.

I suspect this will also help in debugging the issue further a lot.