Remove outputPath fallback from plugin code
webpack will even refuse to run if compiler.outputPath is not set, so it
makes no sense to add CLI fallback logic more suitable for CLI usage
(process.cwd()) to the plugin code.

It should not matter at all what the current working directory is when
the code runs through the plugin code.